refactor: move generate_slug to logic
