+++ eav/fields.py
@@ -1,25 +1,10 @@
 import re
-import secrets
-import string
 
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 
 from eav.forms import CSVFormField
-
-
-def generate_slug(name: str) -> str:
-    """Generates a valid slug based on ``name``."""
-    slug = slugify(name, allow_unicode=False)
-
-    if not slug:
-        # Fallback to ensure a slug is always generated by using a random one
-        chars = string.ascii_lowercase + string.digits
-        randstr = ''.join(secrets.choice(chars) for _ in range(8))
-        slug = 'rand-{0}'.format(randstr)
-    return slug.encode('utf-8', 'surrogateescape').decode()
 
 
 class EavSlugField(models.SlugField):
